+++ src/main/java/ru/alfabank/dmpr/model/ob/ObReportSummaryQualityFilter.java
@@ -19,9 +19,9 @@ public class ObReportSummaryQualityFilter {
     public String direction;
     public Long summary;
     public Long successfully;
-    public Long factkpi;
-    public Long goatkpi;
-    public Long executekpi;
+    public Double factkpi;
+    public Double goatkpi;
+    public Double executekpi;
     public String descript;
     public String duodr_flag;
 }
